+++ src/main/java/de/dm/intellij/liferay/language/TemplateVariableJsonSchema_2_0_Processor.java
@@ -1,0 +1,37 @@
+package de.dm.intellij.liferay.language;
+
+import com.intellij.json.psi.JsonFile;
+import com.intellij.json.psi.JsonObject;
+import com.intellij.json.psi.JsonProperty;
+import com.intellij.psi.PsiFile;
+import de.dm.intellij.liferay.util.LiferayFileUtil;
+
+import java.util.List;
+
+public class TemplateVariableJsonSchema_2_0_Processor implements TemplateVariableJsonSchemaProcessor {
+
+    @Override
+    public boolean isSuitable(JsonFile jsonFile) {
+        String definitionSchemaVersion = LiferayFileUtil.getJournalStructureJsonFileDefinitionSchemaVersion(jsonFile);
+
+        return ("2.0".equals(definitionSchemaVersion));
+    }
+
+    public String getFieldsDefinitionPropertyName() {
+        return "fields";
+    }
+
+    public JsonProperty getFieldReferenceProperty(JsonObject jsonObject) {
+        return jsonObject.findProperty("fieldReference");
+    }
+
+    @Override
+    public JsonProperty getTypeReferenceProperty(JsonObject jsonObject) {
+        return jsonObject.findProperty("type");
+    }
+
+    @Override
+    public JsonProperty getNestedFieldsReferenceProperty(JsonObject jsonObject) {
+        return jsonObject.findProperty("nestedFields");
+    }
+}