+++ src/test/java/de/dm/intellij/liferay/language/freemarker/LiferayFtlVariableProviderTest.java
@@ -61,6 +61,14 @@ public class LiferayFtlVariableProviderTest extends LightJavaCodeInsightFixtureT
         assertFalse("Should not provide name \"simple\", because it is schema version 2.0",  strings.contains("simple"));
     }
 
+    public void testStructureVariablesSimpleJsonDataDefinitionSchema() {
+        myFixture.configureByFiles("WEB-INF/src/resources-importer/journal/templates/test_data_definition/simple.ftl", "WEB-INF/src/resources-importer/journal/structures/test_data_definition.json");
+        myFixture.complete(CompletionType.BASIC, 1);
+        List<String> strings = myFixture.getLookupElementStrings();
+        assertTrue("Should provide fieldReference \"myReference\" for structure based on data definition format", strings.contains("myReference"));
+        assertFalse("Should not provide name \"simple\" for structure based on data definition format",  strings.contains("simple"));
+    }
+
     public void testStructureVariablesSimpleJsonTemplateNode() {
         myFixture.configureByFiles("WEB-INF/src/resources-importer/journal/templates/test/simple-data.ftl", "WEB-INF/src/resources-importer/journal/structures/test.json", "com/liferay/portal/kernel/templateparser/TemplateNode.java");
         myFixture.complete(CompletionType.BASIC, 1);